Split off materialView
